+++ CHANGES.md
@@ -6,6 +6,7 @@
 ### Overview
 
 * Started Helix Swarm integration.
+* Minor improvements to the configuration panel.
 * Bug fixes.
 
 ### Details
@@ -14,13 +15,15 @@
     * The beginnings of Helix Swarm server integration is present in the code.
         It checks whether the Perforce server has a Swarm server registered with it,
         and will attempt to make a connection to the registered Swarm servers.
-* Bug fixes.
+* Minor improvements to the configuration panel.
     * Changed (again) the method for looking up the directory path for different
       configurations.  Configurations now can be associated with the VCS roots,
       rather than the project root.  This means that file location configurations
       will mark themselves as the directory where the file is located (or up to the
       VCS roots); future work may include a "global" file location vs. a local
       file location. (#148) 
+    * Joined problems and resolved properties into a single panel.
+* Bug fixes.
     * Fixed the pop-up link text no longer show an invalid character.
     * Added a minor improvement that limits the number of duplicate "enter a password"
       error messages in some circumstances.